(PC-34835)[API] chore: Move signup validation route under `users` resource
Because we actually patch a user, not some kind of validation resource
