+++ api/tests/routes/pro/validate_user_test.py
@@ -11,13 +11,13 @@ class ValidateUserTest:
         token = token_utils.Token.create(
             token_utils.TokenType.SIGNUP_EMAIL_CONFIRMATION, None, user_id=user_offerer.user.id
         )
-        response = client.patch(f"/validate/user/{token.encoded_token}")
+        response = client.patch(f"/users/validate_signup/{token.encoded_token}")
         assert response.status_code == 204
         assert user_offerer.user.isEmailValidated
 
     @pytest.mark.usefixtures("db_session")
     def test_fail_if_unknown_token(self, client):
-        response = client.patch("/validate/user/unknown-token")
+        response = client.patch("/users/validate_signup/unknown-token")
 
         assert response.status_code == 404
         assert response.json["global"] == ["Ce lien est invalide"]
